refactoring: moved utility type definitions into a separate file
