+++ packages/typir/src/features/equality.ts
@@ -7,8 +7,9 @@
 import { assertUnreachable } from 'langium';
 import { Type, isType } from '../graph/type-node.js';
 import { Typir } from '../typir.js';
-import { TypirProblem, checkValueForConflict } from '../utils/utils-type-comparison.js';
+import { checkValueForConflict } from '../utils/utils-type-comparison.js';
 import { RelationshipKind, TypeRelationshipCaching } from './caching.js';
+import { TypirProblem } from '../utils/utils-definitions.js';
 
 export interface TypeEqualityProblem {
     type1: Type;
