+++ packages/typir/src/index.ts
@@ -24,4 +24,5 @@ export * from './kinds/multiplicity-kind.js';
 export * from './kinds/primitive-kind.js';
 export * from './kinds/top-kind.js';
 export * from './utils/utils.js';
+export * from './utils/utils-definitions.js';
 export * from './utils/utils-type-comparison.js';
